# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nebula Discover API",
    "description": "Public-protein discovery: retrieve \u2192 enrich \u2192 gate \u2192 simulate \u2192 rank. Outputs are unvalidated public-protein candidate hypotheses, never validated sensors.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/objectives/compile": {
      "post": {
        "summary": "Compile Endpoint",
        "operationId": "compile_endpoint_api_objectives_compile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectiveSpec"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "post": {
        "summary": "Create Run",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "summary": "Stream Events",
        "description": "Real Server-Sent Events stream of run progress (not a static JSON list).\n\nReplays events so far, then streams new ones until the run is terminal.",
        "operationId": "stream_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/events.json": {
      "get": {
        "summary": "Get Events Json",
        "operationId": "get_events_json_api_runs__run_id__events_json_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunEvent"
                  },
                  "title": "Response Get Events Json Api Runs  Run Id  Events Json Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/cancel": {
      "post": {
        "summary": "Cancel Run",
        "operationId": "cancel_run_api_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}": {
      "get": {
        "summary": "Get Candidate",
        "operationId": "get_candidate_api_candidates__candidate_id__get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}/dossier": {
      "get": {
        "summary": "Get Dossier",
        "operationId": "get_dossier_api_candidates__candidate_id__dossier_get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateDossier"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}/structure": {
      "get": {
        "summary": "Get Structure",
        "description": "A structure source for Mol*. Prefers the experimental cofactor-bound PDB,\nfalls back to the AlphaFold model. `inline_cif` is populated (from cache/fixture\nor a best-effort fetch) so the viewer works offline; otherwise the client loads\n`provider_url` directly (RCSB/AlphaFold both allow browser CORS).",
        "operationId": "get_structure_api_candidates__candidate_id__structure_get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructureResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CompileRequest": {
        "properties": {
          "objective_text": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 10,
            "title": "Objective Text"
          },
          "user_mode": {
            "type": "string",
            "enum": [
              "novice",
              "expert"
            ],
            "title": "User Mode",
            "default": "novice"
          },
          "instrument_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Id"
          },
          "seed": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Seed",
            "default": 1337
          }
        },
        "type": "object",
        "required": [
          "objective_text"
        ],
        "title": "CompileRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunCreated": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "input_fingerprint": {
            "type": "string",
            "title": "Input Fingerprint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "status",
          "input_fingerprint"
        ],
        "title": "RunCreated",
        "description": "Response to POST /api/runs."
      },
      "RunEvent": {
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "from_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "at",
          "to_status",
          "stage"
        ],
        "title": "RunEvent"
      },
      "RunState": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "input_fingerprint": {
            "type": "string",
            "title": "Input Fingerprint"
          },
          "attempt": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Attempt",
            "default": 0
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "default": "queued"
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 1337
          },
          "objective": {
            "$ref": "#/components/schemas/ObjectiveSpec"
          },
          "instrument_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Id"
          },
          "current_stage": {
            "type": "string",
            "title": "Current Stage",
            "default": "queued"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateRecord"
            },
            "type": "array",
            "title": "Candidates"
          },
          "dossiers": {
            "items": {
              "$ref": "#/components/schemas/CandidateDossier"
            },
            "type": "array",
            "title": "Dossiers"
          },
          "discovery_scores": {
            "items": {
              "$ref": "#/components/schemas/DiscoveryScore"
            },
            "type": "array",
            "title": "Discovery Scores"
          },
          "evidence_shortlist": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Shortlist",
            "description": "candidate_ids on the evidence lane, ranked"
          },
          "frontier_experiments": {
            "items": {
              "$ref": "#/components/schemas/FrontierExperiment"
            },
            "type": "array",
            "title": "Frontier Experiments"
          },
          "generative_frontier": {
            "items": {
              "$ref": "#/components/schemas/GenerativePreview"
            },
            "type": "array",
            "title": "Generative Frontier"
          },
          "selected_candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Candidate Id"
          },
          "result_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Ref"
          },
          "provider_calls": {
            "items": {
              "$ref": "#/components/schemas/Provenance"
            },
            "type": "array",
            "title": "Provider Calls"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/RunEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "offline": {
            "type": "boolean",
            "title": "Offline",
            "default": true
          },
          "status_note": {
            "type": "string",
            "const": "diagnostic_only_not_validated",
            "title": "Status Note",
            "default": "diagnostic_only_not_validated"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "input_fingerprint",
          "objective",
          "created_at",
          "updated_at"
        ],
        "title": "RunState"
      },
...
    }
  }
}